Move xvc files one directory higher
