+++ src/rogue/protocols/CMakeLists.txt
@@ -18,7 +18,7 @@ add_subdirectory("rssi")
 add_subdirectory("srp")
 add_subdirectory("udp")
 add_subdirectory("batcher")
-add_subdirectory("xilinx/xvc")
+add_subdirectory("xilinx")
 
 if (DO_EPICS_V3)
    add_subdirectory("epicsV3")
